WDACConfig module version 0.1.4
Code optimization.
Implemented a check to make sure Windows 11 22H2 is being used as it's a requirement for this module to run.
Improved Remove-WDACConfig cmdlet by making the argument completers only show policies that haven't been already deleted, also applied a fix to this cmdlet for when multiple policies with the same name are deployed.