Removed built-in admin account activation
Due to security reasons. removed the ability to set a password and activate built-in administrator account.

https://support.microsoft.com/en-us/topic/kb5020282-account-lockout-available-for-built-in-local-administrators-bce45c4d-f28d-43ad-b6fe-70156cb2dc00

The lockout policy for the built-in administrator account which is disabled by default:

The new lockout behavior only affects network logons, such as RDP attempts. Console logons will still be allowed during the lockout period.