Harden-Windows-Security-Module v0.0.1
This is a new module that provides proper compliance checking and security score. You can run it on your system to see how compliant it is according to the security measures and guidelines of this repository.

The module performs checks using various methods such as Group Policy, Security Policy, PowerShell cmdlet and Registry keys.

https://www.powershellgallery.com/packages/Harden-Windows-Security-Module/