Harden Windows Security Update v2023.8.30
Optional Diagnostic data is no longer automatically turned on by the miscellaneous category unless user chooses to turn on Smart App Control or it's already on. otherwise you have have an option to turn it on, highly recommended to do it because in the evaluation mode it needs to determine if your system is a good candidate for this AI based security.

https://github.com/HotCakeX/Harden-Windows-Security/issues/94