Improved Process Mitigations
Added thorough explanations to each process mitigation in the CSV file, that will explain why they are used.

This approach logically considers each use case of the mitigations and only implements them if either there are enough information about that process that guarantees it will work 100% with the mitigation and also it makes sense to apply that mitigation in terms of security while also considering usability.

You can always find more info about them in here:

https://learn.microsoft.com/en-us/microsoft-365/security/defender-endpoint/exploit-protection-reference

Removed ForceRelocateImages and RequireInfo from all 1st party executables in the process mitigations list.
The reason is because the former is already enabled by default system wide and the latter is only for older programs. They still exist for 3rd party programs such as Adobe Acrobat but for 1st party programs released by Microsoft they are removed because the RequireInfo mitigation is only for old programs and 1st party programs do not need it.

Removed EnableExportAddressFilter and EnableExportAddressFilterPlus from some processes that might not be compatible with it.

This mitigation is primarily an issue for applications such as debuggers, sandboxed applications, applications using DRM, or applications that implement anti-debugging technology.

Those processes that used them are likely to fall in those categories so to prevent any possible issues or crashes in the future, removed them from the process mitigations.

Import Address Filtering should ideally be used in conjunction with Export Address filtering in order for it to be effective. If an attacker knows you are using Import Address Filtering without Export Address Filtering, they "clould" use the export method to get the address(s) for their shellcode, and vice versa.

Removed DisableNonSystemFonts from Edge browser process mitigations because it uses DirectWrite instead of GDI and this mitigation is not required for it.

Removed EnableRopSimExec as it only applies to 32-bit applications only. quick Assist and Adobe Acrobat that were using it are 64-bit.

Added Hardware Enforced Shadow Stack Protection Strict mode to Edge browser, also added them for Windows Sandbox processes.