Improved Driver block list task schedule creation
The scheduled task used for automatically updating the Microsoft Recommended Driver Block List on the system is now using the SYSTEM account instead of the current Admin account's username and it's using SID instead of account name.

This fixes the issue caused in rare situations where the PC name is the same as the Username, which shouldn't happen.

Also improves the reliability of the task. If the current system Admin account's name is changed or if it's removed, the task will still operate normally.