Removed unnecessary host messages
Removed some unnecessary write-host messages

Removed "PrintDialog.exe" that was a placeholder for the script and module to remove its process mitigations. it was a graceful removal. If you still want to remove it manually, you can run the "Unprotect-WindowsSecurity -OnlyProcessMitigations" command if you have the module installed.

Increased version number in preparation for the release later today