Improved Unprotect-WindowsSecurity cmdlet
For restoration of security group policies, the cmdlet now only restores settings that were changed by Protect-WindowsSecurity excluding the ones applied by the Microsoft Security Baselines.

This allows for a more surgical and careful restoration of the settings and will prevent any accidental changes to the settings that would otherwise need to be constantly kept in check to make sure they are correct.