Improved Harden Windows Security module
The required PowerShell module is now the latest version which is 7.4.0. It has many new features, one of which is the having -ProgressAction common parameter. Using this new common parameter and setting it to SilentlyContinue for Invoke-WebRequest and Invoke-restMethod cmdlets allows for the removal of Invoke-WithoutProgress function since it renders it unnecessary.