Added logic for detecting write-protected drives
Added logic for detecting write-protected drives to be used with Enhanced security level of BitLocker which uses Startup Key as an additional key protector.

Prior to adding the Startup key, we must ensure the drive where the partial key is going to be saved in is writable. This commit adds logic for it.