Improved Hyper-V groups member detection
Using SIDs instead of account names. This makes it more robust and the comparison logic has also been improved.

This change makes everything more inclusive by working in situations where the usernames contain non-English alphabets, lots of spaces and such. Or when the username is the same as the computer name