Added workaround for Controlled Folder Access
The controlled folder access is now properly navigated when using Confirm-SystemCompliance cmdlet.

The new method of BCD NX value verification and detection causes Controlled Folder Access to show notification about pwsh.exe getting blocked so the new change prevents this from happening by dynamically adding pwsh.exe exe to the exclusion list before running the function and then restoring the exclusion list back to exactly how it was at the end of the operation.

This is safely done to ensure that even if user press ctrl + c to prematurely exit the operation or if there is an error, the exclusion list restoration will still happen.