Cleaned up the Registry csv file
Removed the keys marked as Delete. They were there as an automated clean up when each of their categories were run so that the user wouldn't need to manually remove the decommissioned policies.

Also set the SpecialPollInterval to Delete now as it's no longer relevant and will automatically be removed when user runs the Miscellaneous category for automated cleanup.