Updated BitLocker policies
The BitLocker group policies object now applies all of the policies for its category on its own without relying on Microsoft Security baselines. This is required for when user tries to use the BitLocker tab of the GUI to encrypt the drives without applying the protections first. Parts of this change were released in previous updates, this change adds 2 more remaining policies related to disallowing standby modes when BitLocker is enabled.