Improved toast notifications
If toast notifications can't be displayed for any reason such as compatibility issue or other modules conflicting with the dlls, the errors will be suppressed so they won't pollute the logs. It's just a cosmetic feature that has no effect on the actual usage of the module or its policies.