Added Nav bar dot notification for update
The navigation bar will have a small dot with update icon in it whenever a new version is available, attracting user's attention to the update page to take further action.

Added the AutoUpdate user config setting to the PowerShell cmdlets so that the WDACConfig module will be able to utilize it as well.

Removed unnecessary empty lines from some of the PowerShell scripts files.

Updated WDACConfig module file hashes.