Updated AppControl Manager Build Process
The automated pull request now correctly updates the link and also updates the version file.

Updated the pull request name to be better.

Updated Winget download link to always point to the latest release instead of waiting for the aka.ms permalink to be updated, which is slower.

The pull request body is now more readable.