GitHub action workflows update (#390)
Created a custom new action workflow that automatically collects all of the links in markdown files such as the Readme file, and ensures they are up to date and point to the same link mentioned in the markdown document. It generates a nice markdown based job summary too.

Update the AppControl MSIX builder action workflow. The automated pull request now correctly updates the link and also updates the version file.

Updated the pull request name to be better.
Updated Winget download link to always point to the latest release instead of waiting for the aka.ms permalink to be updated, which is slower.
The pull request body is now more readable.
The Virus total uploader action's PowerShell code is now merged with its yaml file.

It no longer posts comment on VT website for uploaded packages because it can lead to unnecessary duplicate comments.
Applied branch protection rules for the Main branch of this repository.