WDACConfig v0.4.8.2 (#393)
This is a small update and as a result only updates the last digit of the version which is the minor version number. It contains the changes listed below:

Using SHA2-512 instead of SHA3-512 for hashing module files in order to be compatible with Windows 11 build 23H2.

Fixed an issue where if you used the Set-commonWDACConfig function to set a user config and you didn't specify the -AutoUpdate parameter, the AutoUpdate property in user config JSON file would be set to false if it was true due to how PowerShell treats boolean parameters by default. That would lead to auto check for update to not happen unless you used the Set-commonWDACConfig -AutoUpdate $true command to turn it on again. It's been taken care of in this version. If that's the case for you and if you desire, please use the Set-commonWDACConfig -AutoUpdate $true to enable auto update again.