Update app.manifest
Added a new section to the Update page that will be displayed when a new version is available. It will display a link to the GitHub releases to show you the release notes.

Added a new option to the Update page that will offer hardened updating procedure. When this option is enabled, the temporary private key of the on-device generated certificate will be linked to the user's account, requiring confirmation of prompts before it can be used for signing.

Added link to X in the about page's links section.

AppControl Manager is now long path aware, it can use file paths longer than 260 characters.