AppControl Manager v1.5.0.0 (#413)
🎉 Added a new page that lets you easily build a code signing certificate to be used for AppControl policy signing. It offers various options to configure the certificate such as key size, common name and validity.

🎉 Added a new page to create Supplemental policies. With this update, you can create supplemental policies by scanning files and folders, or by selecting certificates. More methods will be added in the future such as installed packaged apps and file paths with wildcards. Closes [Bug]: New-SupplementalWDACConfig -Certificates throws XML error "The element 'AllowedSigners' in namespace 'urn:schemas-microsoft-com:sipolicy' has incomplete content" #412

🎉 Added a new page that lets you select XML files and CIP binary files to deploy on the system.

Added a new section to the Update page that will be displayed when a new version is available. It will display a link to the GitHub releases to show you the release notes.

Added a new option to the Update page that will offer hardened updating procedure. When this option is enabled, the temporary private key of the on-device generated certificate will be linked to the user's account, requiring confirmation of prompts before it can be used for signing.

🎉 You can now select a MSIX file path in the Update page to use as the update source instead of downloaded it automatically. You could use this new section if you already downloaded the MSIX package from GitHub or you've built the app yourself from the source code.

AppControl Manager is now long path aware, it can use file paths longer than 260 characters.

Improved the user experience when interacting with settings cards UI elements.

When trying to enter a certificate common name in the settings page, the text box now automatically suggests you all the available common names to choose from. You no longer need to manually type anything. There is also a refresh button that fetches the latest changes to the certificates if you need to update the suggestions.

The system information page will now let you remove supplemental policies from the system with 1 click/tap of a button, just like unsigned policies.

🎉 The app now remembers the window size (including maximized state) when closed, ensuring it reopens with the same size/state next time. Closes [Bug]: AppControl Manager doesn't remember window size #414

Added a notice to the settings page -> Appearance section about Windows animations being disabled: [Bug]: Scroll bars stop working in System Information #415

Set the default path for file and folder pickers to the "WDACConfig" folder in Program Files, instead of the Documents folder. This can reduce the number of clicks user have to use to browse for policies, certificates etc.

Added link to X in the about page's links section.