Harden Windows Security v0.7.0 (#453)
Added Encryption Percentage, Protection Status, Key Protector and Encryption Method properties to the BitLocker tab's Backup section. Those properties are now displayed in the data grid for each drive and will be included in the backup file that you create. This is very useful when you need to view detailed info about the BitLocker protected drives on your system.

Made Audit policy checks available for all System cultures instead of only supporting English-US. This is for the compliance checking feature.

Improved buttons and their positions in BitLocker and Exclusions tabs.

Added a short description to the Exclusions tab.

Slightly improved the performance and speed of compliance checking.

Made lots of performance, quality and security related improvements to the code base.

Fixed this issue -> [Bug]: Small bug in WindowsFeatureChecker.cs #449

Added Long path support policy to the Miscellaneous Category's Intune JSON configuration.

Added the following 3 new policies to the User Account Control Intune JSON configuration:

Behavior Of The Elevation Prompt For Administrator Protection: Prompt for credentials on the secure desktop
Type Of Admin Approval Mode: Admin Approval Mode with Administrator protection
Use Admin Approval Mode: Enabled
Changed this policy in the User Account Control Intune JSON configuration:

Changed this from automatically Deny to "Prompt for credentials on the secure desktop": Behavior Of The Elevation Prompt For Standard Users Prompt for credentials on the secure desktop
Updated the required PowerShell version from 7.4.4 to 7.4.5. The latest available version is 7.4.6 at the moment, which was released over a month ago.