Improving file enumeration in AppControl Manager (#470)
Separated wildcard detection from the rest of the extensions. This improves the extension checks in the tight loops and speeds up the overall enumeration process.

Moved the default App Control supported extensions HashSet outside of the method as static var so it won't be redefined every time the method is called.