Configured default app settings (#475)
The Sidebar's auto-assignment feature is now enabled by default to streamline user interactions.

The check for updates at app startup is now on by default. It simply checks to see if a new version of the app is available and informs the user if there is by showing a small dot on the update page's icon.

Both changes only apply to new app installations. If you've already toggled their buttons off then they remain off.

A border has been added to the error dialog box, styled to match the current Windows accent color.