Implemented more guardrails for signed scenarios
The app won't allow the wrong certificate or common name to be used during signed policy deployment, re-deployment or removal. Such possible user accidents are caught very early on and communicated to the user with proper and clear messages so user can fix the mistake quickly.

The content dialogs that ask for user input for signing scenarios have better visuals and the focus is by default on the Verify button in them which makes it easier and clearer what needs to be done. It also means you can press enter quickly to confirm the actions without using mouse.

Improved DataGrid experience when removing items.

Bumped version from 1.8.0.0 to 1.8.1.0