Implemented online file reputation verification (#507)
Using Microsoft Defender, queries a file's reputation based on either the Smart App Control or SmartScreen, depending on whichever is in control. It doesn't need Admin privileges. It's in a new dedicated tab available in the GUI. Simply browse for a file and detect its reputation and some other advanced details.

Added description texts to the top of ASR rules, Optional features | Apps and the new tab for file reputations.