Improved dependency compatiblity (#514)
The module will not detect if there was a problem loading its dlls and if it was, reloads PowerShell with -NoProfile switch to fix it.

Also a new message will be shown when installing the AppControl Manager and there is an incompatible policy on the system.

Fixes this issue: https://github.com/HotCakeX/Harden-Windows-Security/issues/513