Added support for catalog signed files in local file scans in the AppControl Manager (#533)
Added support for catalog signed files in local file scans. This is a prerequisite for a successful local driver scans since the majority of them are catalog signed. This also improves the overall accuracy of all file scans in the AppControl Manager.

Added support for catalog signed files to the View File Certificates page.