Final removal of WDACConfig
After months of prior notice as well as implementing every feature of the WDACConfig module in the new AppControl Manager and so much more, WDACConfig is fully deprecated. Full speed development continues for AppControl Manager and the Harden Windows Security.

The new applications are built on C# and are so much faster and high performant.