Switch to .NET 9
The Harden Windows Security now uses .NET 9. That means new appearance that is modern, better code, removal of all custom UI elements that belonged to the old WPF designs, faster startup time, support for light/dark theme in the OS, support for accent colors in the OS and so much more.