Further modernization of Harden Windows Security (#572)
Updated the link to the Microsoft 365 apps security baselines to the latest version, 24H2. Previous version was 2306.

Added a new design for when an error occurs in the app

You no longer need to use PowerShell to copy the logs and no error is propagated there. Complete detail of the error is presented to you in the dialog that you see, and with 1 press of a button you can copy it to clipboard and report it on GitHub if you want. This is of course a rare occurrence, but this feature is there whenever it's necessary.
error screenshot
Added support for running the module in Windows Server. You can use all of the features of the Harden Windows Security module in Windows Server 2025 to harden it. This is the Phase 1 of completing this roadmap item.

Applied more optimizations to the code.

Updated Readme with info regarding the new Edge policeis.

Updated the version number file.

Update the required Microsoft DLLs.

Removed the emoji text arts that appear at the end of the compliance check in the CLI experience.

Improved the text colors in the Protect cmdlet in the CLI experience.