Hardening the build process with isolation (#590)
Hardening the build process with isolation. That means multiple jobs/runners/VMs are employed to perform different actions. That also means permissions are assigned based on the least privilege access principle.

The build process now complies with the SLSA L3 security level.

Switched to Windows Server 2025 for the runners, eliminating lots of legacy and old software preinstalled in the runner image and includes all the new security features and legacy feature deprecations in the new Windows Server.