Improving the bootstrapper script (#593)
Removed the old PowerShell link because PowerShell team don't like updating and providing the MSIXBundle package in their releases. Read more about it here: PowerShell/PowerShell#24872

The old link in the bootstrapper script is useless since it points to an old PowerShell version that isn't even supported by the Harden Windows Security module. Winget is the only reliable way to install PowerShell or any apps in general in Windows and if you can't use it for some reason, then you can go ahead and comment in that issue i raised in the PowerShell repository.

Also removed the -MSIXPath parameter from the AppControl function since MSIXBundle parameter is in use now and it can deploy MSIX files too.