Improving multilingual feature of AppControl Manager (#597)
Made lots of changes to the XAML to separate the strings from the code to make it easier to translate the AppControl Manager into other languages.

More changes like this will be made in the future to make the AppControl Manager completely and easily translatable to any other languages.

Improved a log message to be more accurate.