Code refactor and improved resource file usage
Refactor lots of code due to the custom serialization/deserialization logic implemented in the version 1.9.0.0 of the AppControl Manager.

Moved lots of strings to the resource file to improve the ability of the app to be translated into other languages.

Improved the Configure Policy Rule Options page, it now shows details and description of each rule option, helping you to decide which one is suitable for your usecase.