AppControl Manager v.1.9.2.0 (#617)
The User Configurations directory name which has been WDACConfig in the Program Files directory is now changed to AppControl Manager to better reflect the product name. All of the files and directories in the previous directory will be automatically moved to the new directory upon application start. All of the paths in the User Configurations JSON file will be updated to reflect the new directory. The old directory will be removed at the end. This process only happens 1 time, is very fast and requires no user interaction.

WDACConfig was the name of a now deprecated PowerShell module and the directory name was a remnant of that which has been kept unchanged for the sake of interoperability that is no longer the case.
Added a new feature to the Deploy App Control Policies page that lets you select one or more XML files and convert them to CIP files without deploying them.

Added a new feature to the Deploy App Control Policies page that lets you create signed CIP files without deploying them. It will essentially create signed CIP files for you that you can then use to manually deploy them to any other system.

Bumped version to 1.9.2.0.

Improved logging across the app when an error occurs.

Fixed an issue with Allow New Apps' sub-pages where data wouldn't be populated if you were in one of the pages with ListViews and you'd have to switch to the main page and then back to view the data. This was a regression in 1.9.1.0.

Improved the column sorting and naming in pages with ListViews that show data from event logs. They now show more relevant data.

Improved the XML deserialization logic.

Adjusted the locations of all the Guide buttons in each page, they now all appear in the same place, improving consistency.

Fixed an issue where if there was an error while a content dialog was open, the error message wouldn't be displayed. Now content dialogs are tracked globally and if an error occurs while one of them is open, they will be closed for the error message to be displayed in a dedicated content dialog.