Improved List Views and UI/UX in AppControl Manager (#623)
Added keyboard accelerator: CTRL + C for the pages that host a ListView so you can use that shortcut to copy the selected item(s) to the clipboard.

Improved the selection experience in ListViews by making right-click on a row to select that row and then open the context menu.

The settings cards in the Settings page are now interactive and can be clicked/touched to perform the main action in the card.

Added the delete option to the context menu of the ListViews that require it.

Added info bars to the base policy creation page so you can see the live status of policy creation/deployment.