Update AllowNewAppsStart.xaml.cs
All CIP files generated for supplemental and deny policies have the same file name that you select as policy name, making it easier to identify them in the user configurations directory. Previously the CIP files would have the ID (GUID format) which made it hard to recognize which XML or policy they belonged to.

When creating Supplemental or Deny policies, if you choose to deploy them, only the XML policy file will exist in the AppControl Manager directory in Program Files, but if you do not toggle the Deploy button, then the CIP file will also exist in the AppControl Manager directory. This makes it easier for you to use the CIP file on another system. Both the XML and CIP files will have the same name, easy to recognize, and it's the same name you select for the policy.

Made the same change to the Allow New Apps page at the final step (Step 3) when you create to deploy the supplemental policy.

Improved the ListViews for Supplemental and Deny policy creations. The total logs/files are now updated when user is on those pages.