Added new Setting behavior to center list views automatically (#628)
This is a new option in the app settings that is off by default. It is an accessibility setting and also makes it easier to interact with the List Views across the app. Works with touch, mouse and keyboard.

It is added under a new settings category called "Behavior". More customizations will be added under this category in the future. This option, just like others, is remembered between app sessons so you only need to configure it once.