multiple changes
Added progress ring to the auto driver detection in the Create Supplemental Policy page.

Security catalogs scan results are now cached, improving performance and speed of various features that rely on it. Also exposed a new option in the Settings page that lets you turn off/on this feature. It is enabled by default.

more performance improvements