Explicitly defining instruction set, improving the build process and other improvements (#636)
Set instruction set for X64 and ARM64 builds, unlocking new hardware features that AppControl Manager can use.

https://en.wikipedia.org/wiki/ARM_architecture_family
https://en.wikipedia.org/wiki/X86-64#Microarchitecture_levels
https://github.com/dotnet/runtime/blob/main/src/coreclr/nativeaot/docs/optimizing.md
Added lots of useful comments to various sections that provide additional information about the changes added in this PR.

Improved the EditorConfig file.

Bumped the app version to 1.9.6.0 in preparation for the next release.

Updated the build instructions to update the workloads.

https://learn.microsoft.com/en-us/dotnet/core/tools/dotnet-workload-sets