Improved the merge operation in AppControl Manager  (#638)
Improved the merge operation in AppControl Manager.

It is now more accurate, supports FilePath rules and is more high performance because of the multi-threading features implemented.

For the project, reduced the overall package size by removing stack trace generation which also allows for method folding feature in the ILC. Comments are included in the project file for this change.

Removed all unnecessary string interpolations.