Multiple Changes (#639)
Updated Microsoft.IdentityModel.Abstractions dependency to the latest version.

Removed more unnecessary string interpolations.

Added the file name to the FriendlyName field of the Hash rules. Previously the full path of the file would be there which wasn't desired according to user feedback, after that I removed any identifying information completely, but in this version the file name is added to the FriendlyName to provide a minimum information about what each hash rule is for.

Improved performance of various components by changing classes to structs.