Implemented Policy Editor in the AppControl Manager (#645)
The Policy Editor is a highly anticipated feature currently being tracked in this issue. It eliminates the need for manual editing of App Control policies using text editors.

With Policy Editor integrated into the AppControl Manager, users now have access to an intuitive, user-friendly interface designed to streamline policy modifications, making every aspect of policy management more efficient and accessible.

This is only the phase 1 of implementing the Policy Editor feature. Phase 2 will integrate it deeply into the app and will possibly include an option in the Windows 11 context-menu (right-click) where you will be able to select an XML file and open it in the AppControl Manager instantly.

If you have more feedback regarding this feature, feel free to open news issues or comment down below.


Also in this PR:

Bumped the app version to 1.9.7.0
Updated the dependencies to the latest versions, including the .NET SDK and WinAppSDK.