Added clear button to the Policy Editor and created new documents for it (#648)
Added documents for the Policy Editor, can be found here: https://github.com/HotCakeX/Harden-Windows-Security/wiki/PolicyEditor

Added a new button to the policy editor page to clear all of the data displayed on the UI.

Fixed the main window not being draggable after switching to the Policy Editor page.

Improved the search performance in all of the List View elements.