New Admin prompt checkbox, simulation and merge improvements, code refactoring in the AppControl Manager (#665)
Added a checkbox to the elevation prompt when accessing admin-required pages from unelevated session to enable auto elevation at startup.

Added a new option to the settings page that can also control the startup elevation behavior. When this option is enabled, AppControl Manager will prompt for elevation by default at startup just like before version 1.9.8.0. The difference now is that this behavior can be controlled by the user and is no longer static.

Bumped the app version to 1.9.9.0.

Improved the accuracy of the Merge and Simulation operations.

Lots of general code improvements such as type internalization and refactoring.

The update page and its associated logic no longer work when the app is installed from the Microsoft Store (for future preparations).

Streamlined the app settings management internally.