Added new helpful UI elements to the AppControl Manager
Added a new message bar to the Create Policy From Event Logs to show real time messages about the currently running processes.

Added a new button called Only Use Selected Items to the Create Policy From Event Logs and Create Policy From MDE Advanced Hunting pages. The following is the description of the button:

Only include the selected items in the supplemental policy. You can select one or more items in the ListView and if this option is enabled, only those will be included in the supplemental policy. If this is disabled, everything will be included in the supplemental policy. If you use this option and nothing is selected in the ListView, everything will be included in the supplemental policy.
Added a Progress Ring to the Policy Editor page to show the progress when a policy is being loaded.

Added additional helpful insights to many UI elements for accessibility purposes so tools such as screen readers will be able to use them to generate proper feedback for people that need them.

Added more messages to the message bar in the Create Policy From MDE Advanced Hunting page. They offer more information about the current operation and its progress.

Added a button called Open In The Policy Editor to the following pages:

Create AppControl Policy

Allow New Apps

Create Policy From Event Logs

Create Policy From MDE Advanced Hunting

This feature enables users to seamlessly open newly created policies directly in the Policy Editor, offering a detailed overview of the policy, its associated rules, and the ability to refine and customize it further. By streamlining this workflow, the feature significantly reduces friction and enhances the overall user experience. The button only appears at the end of the operation on the message bar.